+++ autumn/src/app.rs
@@ -521,9 +521,9 @@ impl AppBuilder {
             started_at: std::time::Instant::now(),
             health_detailed: config.health.detailed,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new(&config.log.level),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::from_config(&config),
+            log_levels: crate::diagnostics::LogLevels::new(&config.log.level),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::from_config(&config),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -652,9 +652,9 @@ impl AppBuilder {
             started_at: std::time::Instant::now(),
             health_detailed: config.health.detailed,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new(&config.log.level),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::from_config(&config),
+            log_levels: crate::diagnostics::LogLevels::new(&config.log.level),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::from_config(&config),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -1461,9 +1461,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -1550,9 +1550,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -1631,9 +1631,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -1679,9 +1679,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -1837,9 +1837,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -2006,9 +2006,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -2141,9 +2141,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -2174,9 +2174,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             #[cfg(feature = "ws")]
             channels: crate::channels::Channels::new(32),
             #[cfg(feature = "ws")]
@@ -2366,9 +2366,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             channels: crate::channels::Channels::new(32),
             shutdown: tokio_util::sync::CancellationToken::new(),
         };
@@ -2418,9 +2418,9 @@ mod tests {
             started_at: std::time::Instant::now(),
             health_detailed: true,
             metrics: crate::middleware::MetricsCollector::new(),
-            log_levels: crate::actuator::LogLevels::new("info"),
-            task_registry: crate::actuator::TaskRegistry::new(),
-            config_props: crate::actuator::ConfigProperties::default(),
+            log_levels: crate::diagnostics::LogLevels::new("info"),
+            task_registry: crate::diagnostics::TaskRegistry::new(),
+            config_props: crate::diagnostics::ConfigProperties::default(),
             channels: crate::channels::Channels::new(32),
             shutdown: tokio_util::sync::CancellationToken::new(),
         };
