+++ autumn/src/lib.rs
@@ -74,6 +74,7 @@ pub mod channels;
 pub mod config;
 #[cfg(feature = "db")]
 pub mod db;
+pub mod diagnostics;
 pub mod error;
 pub mod error_pages;
 pub mod extract;
